Add kmakemodule, which generates .module files from XML descriptions of modules.
The reason for having this rather than writing the .module files by hand is
that they contain a significant amount of boilerplate. For example, kroc and
the slinker currently have different ways of specifying FFI libraries to use.
This script can generate it all from a simpler description, and means that if
we want to add extra stuff to the .module files in the future (for example, for
nocc or tock), we only need to change the script.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@3073 ee8ce1bd-81f8-0310-a706-c2403d3965fe