Clean out some code in dmem.c.  Fix a minor bug in global structure initialisation. Add statistics interface, intended usage is as follows:
  #include <ccsp.h>

  void some_function (void)
  {
    ccsp_stats_t stats;
    int active_procs;

    if (ccsp_get_stats (CCSP_STATS_VERSION, &stats) < 0) {
      fprintf (stderr, "Error obtaining stats...");
      return;
    }

    active_procs = (stats.proc_start - stats.proc_end) + (stats.startp - stats.endp);
    fprintf (stderr, "%d active processes\n", active_procs);
  }

No locking or synchronisation is applied so it's possible for the statistics to be skewed.
I haven't yet check the performance hit for maintaining the stats, if it is high then they may become a compile time option; however, the C function will continue to be exported to maintain binary compatibility.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4267 ee8ce1bd-81f8-0310-a706-c2403d3965fe