This adds pre-defines for binding mobiles:
  PROC BIND.MOBILE (MOBILE.ANY mobile, VAL INT address)
  PROC BIND.MOBILE.HW (MOBILE.ANY mobile, VAL INT physical.address)

These are intended for low-level work, not for general user consumption.

Say we have a piece of system memory that we want to pass around our program as a MOBILE, the suggested code sequence is as followed:
  
  MOBILE []BYTE x:
  INT mem.addr, mem.len:
  SEQ
    mem.addr := ... -- can be variable
    mem.len := ... -- can be variable, unlike PLACED []BYTE

    x := MOBILE EMPTY [mem.len]BYTE
    BIND.MOBILE (x, mem.addr)

There is no support for unbinding, binding is a destructive process, although re-binding a mobile to the null pointer (0), might be used to generate errors on reuse.

You'll also notice I've added a MOBILE.ANY type, this is only supported for definition of pre-defined or external processes.  It does mean one can do the following:

#PRAGMA EXTERNAL "PROC C.do.mobile.stuff (MOBILE.ANY mobile) = 0"

To declare an external process taking any mobile.  Again, this is very specialist and should be used with caution.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4297 ee8ce1bd-81f8-0310-a706-c2403d3965fe