Fix CONTRSPLIT handling so that it maybe followed instructions other than CJ.  Since there is no formal specification for ETC I don't know what is legal, but for now allow only FPCHKERR.  Fixes #59.
The problem is this unearthed bad merging of register contrains by the colouring code.  Constrains were merged without respect to the fact that other virtual registers may constrain to the same physical in the intervening code.  This has been fixed and colouring works again for these cases.

There is also some additional cleaning of the FPCHKERR code generation for i386.

Passes all cgtests; I'm reasonably confident it is correct.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4394 ee8ce1bd-81f8-0310-a706-c2403d3965fe