This implements tighter/cleaner kernel call integration with GCC using register parameters on x86, the mechanism should be portable to other platforms.  It significantly reduces the assembly code burden, and marks most of the way forward to building position independent shared libraries.
To aid PIC, EBP and EBX are now treated as respectfully as possible inside the kernel and tranx86 output.  Kernel calls all use the 'CALL' instruction, instead of a range of different jumps with compile time addressing.

On x86, register parameters are used to pass the first parameter of the call in EAX, the schedule pointer in EDX and the workspace pointer in ECX.  Any additional parameters are stored to parameter offsets in the scheduler data structure.  A set of macros abstract this within the kernel code, so if the mechanism changes (different platform) only the macros need change.  Exiting the kernel is still performed via a set of assembly macros, and will probably continue to be this way as I can't see a good way to manipulate the return address and restore the workspace and scheduler pointers as the occam code expects.

tranx86 now never uses the EBX register, EAX, EDX, ECX and EDI are used for normal code generation, this means that some operations requiring byte access must be bound to EAX, EDX, or ECX, as EDI only supports 32 or 16 bit accesses.  This work is now all done.

The generated code pass all cgtests, and can run hereticc.  This does however not preclude the possibility, of some undetected bugs.  CIF is known to not work, it is currently only 40% working, and will be restored to full functionality shortly, when I've found a nicer way of re-implementing some of its functionality.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4460 ee8ce1bd-81f8-0310-a706-c2403d3965fe