Fix the tree transformation for BYTESIN so that it works properly with mobile arrays.
This used to look for a -1 return from known_bytesin to tell whether the object
in question was mobile. Since Carl fixed known_bytesin to understand mobiles in
[4395]/[4396], there's no need to do that any more; it can just strip the
S_MOBILE off and proceed as if it were a normal array.

(It's probably broken for mobile arrays of regular arrays or similar, but it's
in good company there...)

Fixes #77.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4479 ee8ce1bd-81f8-0310-a706-c2403d3965fe