This is very rough and I intend to do a fair bit of cleaning on it, but committing as a working example.
This firmware receives a .SRV from the UART0, then uses a special FFI call to run it in the user context.  A top-level set of channels (in? and out!) is provided.  The firmware moves data from the UART to and from the user program.

I've included an "id" program, which I've been using for testing, it simply copies input to output.  By uploading it I create the following circuit:

 UART0 => firmware => user
                       |
 UART0 <= firmware <= user

I can stream data through this at around 45kb/sec.  Which is pretty good considering everything that's going on (only 10-15kb/sec drop from the firmware doing it directly).

As aside I had to fix my UART code as I was accidently twiddling the wrong wire for flow control, but all is now good.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4602 ee8ce1bd-81f8-0310-a706-c2403d3965fe