Make the disconnected channel model more consistent now I'm happy about how to deal with it. Disconnected channels are now always ready even in ALTs.  This makes them somewhat like floating disconnected wires.
I've reworked the SRV-1 firmware to take account of this and simplified code appropriately, although we now test disconnection on the output stream so that the user can send 0 bytes without trouble.  Overall the direct output path from user code to the UART0 gets about 50-55kb/sec throughput.  For higher-speeds the a blit FFI function is still provided.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4647 ee8ce1bd-81f8-0310-a706-c2403d3965fe