Clean up the JPEG code further to remove the use of undefined C operations, like the following:
  *Y2_Ptr++ = *(Y2_Ptr - 1)

Where it isn't defined if the right side is evaluated before the left.

Also push the encoder structure off the stack.  The encoder uses lots of static buffers anyway, so the state structure can also be static.  I assume this is legacy from when the encoder was once properly re-entrant, but as far I know the SRV-1 version never has been.

I've spotted and marked one possible bug in the 4:2:2 reader, but I haven't fixed it as it may be the intended behaviour.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4781 ee8ce1bd-81f8-0310-a706-c2403d3965fe