Add (untested) RESIZE.MOBILE.ARRAY.1D.
Think of it like:
  PROC RESIZE.MOBILE.ARRAY.1D (RESULT MOBILE.ANY m, VAL INT new.dimension)

I've also renamed the mtdef.h in occ21 to mobile_types.h to keep it consistent with the rest of the copies.  Added a new MT_RESIZE instruction, which at the moment only handles array sizes, but is extensible to handle just about any type of resize operation.

There is a cgtest for RESIZE.MOBILE.ARRAY.1D, which I'm now going to start poking to check that it works.  If I get time I may try and add a generic RESIZE.MOBILE.ARRAY function which takes an array of dimensions.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4783 ee8ce1bd-81f8-0310-a706-c2403d3965fe