Enable 32KiB of spare data SRAM as a data cache for the SDRAM.  This should enable most user code to execute as if it were in the much faster SRAM even though it is backed in the SDRAM.  Generally this should give a significant performance boost, although I'm not going to do any in-depth benchmarks now.
The data SRAM is divided into two 32KiB banks.  The upper-half of each bank can be configured to act as cache.  To enable this functionality we configure a static page-table using 9 of the 16 data cache cacheability protection lookaside buffer (DCPLB) registers.  These entries grant read-write access to the SRAM and SDRAM, while marking the SRAM as uncachable and the SDRAM cachable.  We configure all the cache as write-back, which means no writes go out to the SDRAM unless we run out of cache-lines.  For reference Blackfin cache-lines are 32 bytes in length.

The only slight complexity (assuming you consider what I've described so far as being simple), is that the DMA controller interacts directly with the SDRAM bypassing the cache.  Therefore to maintain synchronisation we must flush the cache-lines for any buffers we DMA into (i.e. read camera frames into).  This is relatively straight forward, and costs a couple of thousand CPU cycles per-frame on small (320x256) images, to a couple of hundred-thousand cycles on the largest frame sizes.

In practice I really can't see any slow down from the cache flushes.  Also note we only need to flush the cache when we allocate a DMA buffer, which means we only pay a hit when we take a frame out of the DMA ring.  This means that if you run the camera in stream mode in the background and take the occasional frame then you only pay for the frames you take.

Finally because the data SRAM is now half the size things are a bit tighter, although this causes no immediate problems.  At the moment there is still 14KiB of free data SRAM, but it is spread between the two banks.  If we add any large static data to the firmware some manual compiler hinting maybe required to put it in a different bank.  We also alway have the option of reclaiming 16KiB of the cache for use again as SRAM.  That said, I think given user code runs out of the SDRAM the more caching we can do the better.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4804 ee8ce1bd-81f8-0310-a706-c2403d3965fe