Just commit some example code (disabled) to use the performance monitors to see whether the cache is being used.  The short answer is yes.
For the SRV1 protocol emulator capturing images at 320x256, all the cache does is pull in data from SDRAM, it almost never flushes any memory.  This is what one would expect as the JPEG encoded frames are small enough to fit in the cache.  For 1280x1024 frames I see a small flow of cache-line flushes back to the main memory, but the predominately reads in.  When no image captures being performed there is no cache-line traffic.  Which is what we would hope; it means the user code is running from the SRAM.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4806 ee8ce1bd-81f8-0310-a706-c2403d3965fe