Note: This code is not yet intended for general consumption, I am committing it to make sure there is a copy in the public domain.
The plinker directory contains a complete TCE to TEncode TBC assembler and linker, with library, FFI and optimal-prefixing support.  It is meant to be a short-lived exploration on the road to a Tock backend targeting the TVM and a new registered based TVM.

Why I did this:
  1. I need to understand the complete problem of assembling and linking ETC.
  2. I wanted to test the idea that linking could be done in simpler way and more efficiently than it is at present slinker.

Given (1) and my lack of confidence in Scheme, the hit of writing from scratch in a different language rather than extending and modifying the slinker seemed justified.  The actual linker took me only 3 days to complete, and I have only been committing to these code in brief intervals when not working on other things.

Perl has the advantage that it is widely available, should this code be used as an interim component of the bytecode generation chain for the Transterpreter.  Although I should stress that is not my objective.  This code is, for the time being, a personal exploration.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@4870 ee8ce1bd-81f8-0310-a706-c2403d3965fe