All of the changes are minor, and were just us getting things to the point that we could say "this code does (or does not) run."
As it turns out, it "does not." We spent most of the evening not working, so this last commit is just so Christian can take a look tomorrow while I go to a meeting. Very, very exciting.

Things should just be builable with

ARMGCC=<PATH> make download

which will build everything (including libtvm for ARM) and load it onto the Fluke using lpc21sp (or whatever it is called). 



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5041 ee8ce1bd-81f8-0310-a706-c2403d3965fe