This is a big/odd commit. I will explain.
We began with the Fluke firmware as it ships with the bot. As it turns out, a
fair bit of that firmware is borrowed from RO Software; Bill Knight (I believe
that's the author's name) released some widely used/followed example code for
the LPC210x series of processors. The IPRE group have used this code in their
firmware.

However, they didn't use it verbatim. Instead, there are constants that are
renamed, and changes in places, and some things are hard-coded instead of
computed. Now, this isn't a problem unless you are trying to read the
documentation and make sense of their firmware. You see, by changing the names
of constants (registers, etc.), you can no longer look at a register name and
look it up in the documentation.

This is frustrating.

It seems, even, that they took the RO Software code and did name-changes in it,
meaning that I can't even cleanly use other bits of the ROS example code. I have
to go through and rename at least one of the code-bases. I can either rename to
"broken" names, or I can rename the Fluke firmare, falling out of sync with
their work.

In the end, their code is:

1. A loop that reads commands from a remote host over Bluetooth and
2. A framegrabber.

They need two UARTs and some specialized assembly for working with the camera.
And, kludging in interrupts into their code is less attractive than simply
starting with the ROS codebase that has interrupts (for both UARTs) baked-in. In
fact, you can turn them on and off by commenting or uncommenting one or two
definitions. Likewise, UART0 and UART1 can be included in the code (or excluded)
at compile-time with a single value in a #define.

In short, the RO Software codebase is cleaner, and better organized. It looks
like the Fluke team started a hardware port from scratch, and imported bits
where they found they could. (This is a guess... I have no idea.)

Our strategy, at this point, is going to be to start from the RO Software
codebase, and make as few (zero, if possible) changes to it. Those changes that
we make will be specifically to support integration with the Transterpreter.
Then, we'll bring in the limited amount of code that is actually
Fluke-specific. In this case, that would be:

1. The code to drive the camera, and 
2. The code they download onto the BASIC stamp (or whatever it is) inside of the
Scribbler.

This seems like the cleaner way to go. So, we're going to commit the RO Software
codebase in a state that builds and executes on the Fluke, and then bring in our
Transterpreter bits. We'll then get rid of the Fluke directory, and possibly
rename this one. 

That's the long story of the day. 



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5054 ee8ce1bd-81f8-0310-a706-c2403d3965fe