We have tied into the TVM's scheduler, and it is good.
So, we have to create a new interrupt in the enumeration table, and in the IRQ,
raise that interrupt. This way, the scheduler is aware that the universe has
changed. 

The magic channel needs to store information away when it reschedules. In
particular, we need to make sure we have a pointer to the "channel word." This
gets used when the TVM clears the pending interrupts; the execution context
gets passed in (the firmware context), and we use it and the channel word to
put ourselves back on the queue. 

I think that's the SVN commit message version of "here's how the magic
interrupt-driven channels interact with the scheduler" document, so that Carl
can decide if we're "in the ballpark." (We're coding in the US, so it's a
ballpark, not a cricket pitch.)

Christian is complaining about the length of this commit message, so I end it
here. The next commit will remove all of the debug messages which we inserted.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5065 ee8ce1bd-81f8-0310-a706-c2403d3965fe