This represents the state that we're at with the fluke.
Interrupts (clock + uart incoming) are working, except for some random crashing
on the uart interrupts. We'll debug that more when we can get OpenOCD to talk
to our dev board. There is no JTAG on the fluke :/

Stuff compiles either for the fluke or the dev board, run
  make config target=<FLUKE|OLIMEX>
to switch.

I also renamed the download targets to program (which is consistent with what
the makefile originally used, and download was not the right thing here anyway,
but is what the original makefile we copied was using)



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5081 ee8ce1bd-81f8-0310-a706-c2403d3965fe