Fix the -C comment decoder in tranx86.
This was breaking for long strings, because the comment length is
encoded in tcoff's compressed format, and the decoder I'd written was
oring the bytes back together in the wrong order. This hadn't been a
problem before because the comments were always short enough to be
encoded as a single byte -- but the libGL workaround on MacOS X pushed
the .LDFLAGS comments over 250 bytes!

This also renames tcoff_getul to tcoff_getl, since the type it's
decoding is a signed rather than unsigned long in the other tcoff code.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5409 ee8ce1bd-81f8-0310-a706-c2403d3965fe