Merging most of the changes from 5587:5656 from my tock branch into trunk.
This is mainly changes to get things working with Tock, but there's also one or two fixes in there too.

On the way, I removed nearly all the extra type-specifiers that I'd added to get things compiling in Tock, as hopefully I won't need those soon.

This merge *shouldn't* break any of the existing KRoC setup, but hopefully the build bots will show if it does.



git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5657 ee8ce1bd-81f8-0310-a706-c2403d3965fe