Implement a custom memory interface for the AVR which maps program memory into the top half of the address space.
The AVR is a Harvard-architecture processor, so it has separate address
spaces for data and code. The program memory is typically much bigger
than the RAM (e.g. 32k vs. 2k on the processor we're using), so we want
to be able to put our transputer bytecode into program memory. We do
this by intercepting reads to addresses above 0x8000 and doing a program
memory read instead. All the other memory interface functions stay the
same as the native memory interface, since data addresses are mapped
directly, and we don't ever need to write to program memory.

We provide an additional function in libtvm to map a program memory
address to a Transterpreter address (tvm_addr_from_progmem), and use
this when calling tvm_ectx_install_tlp in the Arduino wrapper.

There is some additional complication for FFI code, since it also needs
to go via the Transterpreter memory interface whenever it's given a
pointer (e.g. serialWrite might have a pointer into program memory).
We include a memcpy_from_tvm function that might go into libtvm at some
point.

This also adds a configure test for the AVR CPU to set the word
length and memory interface automatically, and fixes automatic
word-length setting for various other processors that had a typo in a
variable name.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5877 ee8ce1bd-81f8-0310-a706-c2403d3965fe