Make the Arduino wrapper decode and run a TEncode file from a particular location in program memory, rather than just having an array compiled into it. This means you can just upload a new occam program to the board, rather than needing to upload the Transterpreter too each time.
We can't use the libtvm TEncode decoder directly, since it doesn't know
about program memory. For now, this is a copy-and-paste-and-modify of
the existing code, with all the stuff we're not using stripped out, and
the types and function calls changed to read from program memory. (We
didn't encounter any other problems while doing this, so we think the
existing decoder should work fine on 16-bit platforms that don't have
the program memory weirdness.)

The Makefile has gained an "upload-occam" target -- with a script to
wrap a TEncode (or whatever) file as the limited subset of Intel HEX
that avrdude understands -- and uses the plinker when compiling occam
programs.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5884 ee8ce1bd-81f8-0310-a706-c2403d3965fe