Fix the autoreconf-not-detecting-subdirs problem.
Sleeping on the problem helped: the solution is to actually call
AC_CONFIG_SUBDIRS in our equivalent macro (so autom4te detects it's
being called), but undo its effects that we don't want. This also avoids
duplicating some of AC_CONFIG_SUBDIRS' logic in
OCCAM_CONFIG_SUBDIRS_FOR, which is nice.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5913 ee8ce1bd-81f8-0310-a706-c2403d3965fe