Map an interrupt to a placed channel.
This is an initial experiment that we're going to generalise later, but
it shows how to do this in the simplest case: an external channel
routine for occam programs to use, an interrupt handler, and a
clear_pending_interrupts routine that requeues the process once the read
is complete.

We spent a while getting the waiting_on_interrupts logic correct;
initially we were testing timer_ptr rather than timer_channel, which
doesn't do the right thing since it's only set once the interrupt
handler has run...

For testing, we've set up one of the AVR timers to give us periodic
interrupts; this'll disable some of the PWM outputs for now.


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5916 ee8ce1bd-81f8-0310-a706-c2403d3965fe