On the Mac, avrdude is not happy with the parallel-port programmers. Even though we are not using one, it throws an error when it encounters one where
  type = par

in the config file. While there may be a way to work around this, I have
removed the configuration options for programmers based on a parallel port from
avrdude.conf. In short, if you are not programming your Arduino over the USB
cable, then we aren't going to support it. 


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@5981 ee8ce1bd-81f8-0310-a706-c2403d3965fe