Add $(srcdir) to the source filename in all the invocations of occbuild --program. This means you can build programs in a separate object directory.
(The implicit rules -- i.e. ".occ:" rules -- don't need this, because
they do a VPATH search for the source automatically. There are a few
directories at the moment that could use one of these rules but don't;
it's safe to do so when none of your programs have more than one source
file.)


git-svn-id: https://projects.cs.kent.ac.uk/projects/kroc/svn/kroc/trunk@6023 ee8ce1bd-81f8-0310-a706-c2403d3965fe