Signals now compile and (may) output correct code
relement in syn2.c now returns a new DECL node where it sees SIGNAL on the RHS
of a channel comms. This makes the RHS parse.

chk_gettype_main, now returns S_SIGNAL as being an int, for symmetry with the
size of what is RHS of the communication is (currently an INT)

bytesinscalar returns 4 for things of type S_SIGNAL