Broken DEBUG_MSG
One of the DEBUG_MSGs in this file contains a variable that is not in scope. It
is broken, and should be discarded or fixed. We commented it out, which is not
a better option, but it allowed us to compile and use debugging.

(We have no idea what the out-of-scope variable was used for...)